requirements-dev.txt
bandit[toml]==1.7.9
# Development dependencies for testing, linting, and security
coverage==7.6.1
pre-commit==3.8.0

# Optional: GitHub API for label management
PyGithub==2.4.0
pytest-cov==5.0.0
ruff==0.6.9
safety==3.2.7
